refactor(mcp): split report and resource helpers
