+++ crates/codelens-mcp/src/main.rs
@@ -8,9 +8,11 @@ mod error;
 mod mutation_gate;
 mod prompts;
 mod protocol;
+mod resource_context;
 mod resources;
 mod runtime_types;
 mod server;
+mod session_metrics_payload;
 mod state;
 mod telemetry;
 mod tool_defs;
