+++ crates/codelens-mcp/src/tools/reports.rs
@@ -1,460 +1,8 @@
 use super::report_contract::make_handle_response;
-use super::{AppState, ToolResult, required_string, success_meta};
-use crate::error::CodeLensError;
-use crate::protocol::BackendKind;
+use super::report_utils::{stable_cache_key, strings_from_array};
+use super::{AppState, ToolResult, required_string};
 use serde_json::{Value, json};
 use std::collections::BTreeMap;
-use std::path::Path;
-use std::time::Duration;
-
-fn strings_from_array(value: Option<&Vec<Value>>, field: &str, limit: usize) -> Vec<String> {
-    value
-        .into_iter()
-        .flatten()
-        .take(limit)
-        .filter_map(|entry| {
-            if let Some(text) = entry.as_str() {
-                Some(text.to_owned())
-            } else if let Some(obj) = entry.as_object() {
-                obj.get(field)
-                    .and_then(|v| v.as_str())
-                    .map(ToOwned::to_owned)
-                    .or_else(|| Some(entry.to_string()))
-            } else {
-                Some(entry.to_string())
-            }
-        })
-        .collect()
-}
-
-fn stable_cache_key(tool_name: &str, arguments: &Value, keys: &[&str]) -> Option<String> {
-    let mut fields = BTreeMap::new();
-    for key in keys {
-        if let Some(value) = arguments.get(*key) {
-            if !value.is_null() {
-                fields.insert((*key).to_owned(), value.clone());
-            }
-        }
-    }
-    if fields.is_empty() {
-        None
-    } else {
-        Some(
-            json!({
-                "tool": tool_name,
-                "fields": fields,
-            })
-            .to_string(),
-        )
-    }
-}
-
-fn extract_handle_fields(payload: &Value) -> (Option<String>, Vec<String>) {
-    let analysis_id = payload
-        .get("analysis_id")
-        .and_then(|value| value.as_str())
-        .map(ToOwned::to_owned);
-    let estimated_sections = payload
-        .get("available_sections")
-        .and_then(|value| value.as_array())
-        .map(|items| {
-            items
-                .iter()
-                .filter_map(|item| item.as_str().map(ToOwned::to_owned))
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-    (analysis_id, estimated_sections)
-}
-
-fn run_job_kind(state: &AppState, kind: &str, arguments: &Value) -> ToolResult {
-    match kind {
-        "impact_report" => impact_report(state, arguments),
-        "dead_code_report" => dead_code_report(state, arguments),
-        "refactor_safety_report" => refactor_safety_report(state, arguments),
-        _ => Err(CodeLensError::Validation(format!(
-            "unsupported analysis job kind `{kind}`"
-        ))),
-    }
-}
-
-fn debug_step_delay_ms(arguments: &Value) -> u64 {
-    arguments
-        .get("debug_step_delay_ms")
-        .and_then(|value| value.as_u64())
-        .unwrap_or(0)
-        .min(250)
-}
-
-fn maybe_delay(ms: u64) {
-    if ms > 0 {
-        std::thread::sleep(Duration::from_millis(ms));
-    }
-}
-
-fn estimated_sections_for_kind(kind: &str) -> Vec<String> {
-    match kind {
-        "impact_report" => vec!["impact_rows".to_owned()],
-        "dead_code_report" => vec!["candidates".to_owned(), "raw_dead_code".to_owned()],
-        "refactor_safety_report" => vec![
-            "module_boundary".to_owned(),
-            "symbol_impact".to_owned(),
-            "change_request".to_owned(),
-            "related_tests".to_owned(),
-        ],
-        _ => Vec::new(),
-    }
-}
-
-fn patch_job_file(
-    project_path: &str,
-    job_id: &str,
-    status: Option<&str>,
-    progress: Option<u8>,
-    current_step: Option<Option<String>>,
-    analysis_id: Option<Option<String>>,
-    error: Option<Option<String>>,
-) {
-    let path = Path::new(project_path)
-        .join(".codelens")
-        .join("analysis-cache")
-        .join("jobs")
-        .join(format!("{job_id}.json"));
-    let Ok(bytes) = std::fs::read(&path) else {
-        return;
-    };
-    let Ok(mut job) = serde_json::from_slice::<crate::state::AnalysisJob>(&bytes) else {
-        return;
-    };
-    if let Some(status) = status {
-        job.status = status.to_owned();
-    }
-    if let Some(progress) = progress {
-        job.progress = progress;
-    }
-    if let Some(current_step) = current_step {
-        job.current_step = current_step;
-    }
-    if let Some(analysis_id) = analysis_id {
-        job.analysis_id = analysis_id;
-    }
-    if let Some(error) = error {
-        job.error = error;
-    }
-    job.updated_at_ms = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis() as u64;
-    if let Ok(updated) = serde_json::to_vec_pretty(&job) {
-        let tmp_path = path.with_extension("json.tmp");
-        let _ = std::fs::write(&tmp_path, updated);
-        let _ = std::fs::rename(tmp_path, path);
-    }
-}
-
-fn advance_job_progress(
-    state: &AppState,
-    job_id: &str,
-    progress: u8,
-    current_step: &str,
-    delay_ms: u64,
-) -> Result<bool, String> {
-    if state
-        .get_analysis_job(job_id)
-        .as_ref()
-        .map(|job| job.status.as_str())
-        == Some("cancelled")
-    {
-        return Ok(false);
-    }
-    state
-        .update_analysis_job(
-            job_id,
-            Some("running"),
-            Some(progress),
-            Some(Some(current_step.to_owned())),
-            None,
-            None,
-            None,
-        )
-        .map_err(|error| error.to_string())?;
-    maybe_delay(delay_ms);
-    Ok(true)
-}
-
-fn run_job_kind_with_progress(
-    state: &AppState,
-    job_id: &str,
-    kind: &str,
-    arguments: &Value,
-) -> Result<Value, String> {
-    let delay_ms = debug_step_delay_ms(arguments);
-    match kind {
-        "impact_report" => run_impact_report_job(state, job_id, arguments, delay_ms),
-        "dead_code_report" => run_dead_code_report_job(state, job_id, arguments, delay_ms),
-        "refactor_safety_report" => {
-            run_refactor_safety_report_job(state, job_id, arguments, delay_ms)
-        }
-        _ => run_job_kind(state, kind, arguments)
-            .map(|(payload, _meta)| payload)
-            .map_err(|error| error.to_string()),
-    }
-}
-
-fn run_dead_code_report_job(
-    state: &AppState,
-    job_id: &str,
-    arguments: &Value,
-    delay_ms: u64,
-) -> Result<Value, String> {
-    let scope = arguments
-        .get("scope")
-        .and_then(|v| v.as_str())
-        .unwrap_or(".");
-    let max_results = arguments
-        .get("max_results")
-        .and_then(|v| v.as_u64())
-        .unwrap_or(20);
-    if !advance_job_progress(state, job_id, 20, "scanning dead code candidates", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let dead_code =
-        super::graph::find_dead_code_v2_tool(state, &json!({"max_results": max_results}))
-            .map(|output| output.0)
-            .map_err(|error| error.to_string())?;
-    if !advance_job_progress(state, job_id, 70, "filtering scoped dead code", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let candidates = dead_code
-        .get("dead_code")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_default()
-        .into_iter()
-        .filter(|entry| entry.to_string().contains(scope))
-        .take(10)
-        .collect::<Vec<_>>();
-    let top_findings = strings_from_array(Some(&candidates), "file", 3);
-    let mut sections = BTreeMap::new();
-    sections.insert(
-        "candidates".to_owned(),
-        json!({"scope": scope, "dead_code": candidates}),
-    );
-    sections.insert("raw_dead_code".to_owned(), dead_code);
-    if !advance_job_progress(state, job_id, 90, "writing dead code analysis", delay_ms)? {
-        return Ok(json!({}));
-    }
-    make_handle_response(
-        state,
-        "dead_code_report",
-        stable_cache_key("dead_code_report", arguments, &["scope", "max_results"]),
-        format!("Bounded dead-code audit for scope `{scope}`."),
-        top_findings,
-        0.84,
-        vec!["Validate runtime entry points before deleting candidates".to_owned()],
-        sections,
-        if scope == "." {
-            Vec::new()
-        } else {
-            vec![scope.to_owned()]
-        },
-        None,
-    )
-    .map(|(payload, _meta)| payload)
-    .map_err(|error| error.to_string())
-}
-
-fn run_impact_report_job(
-    state: &AppState,
-    job_id: &str,
-    arguments: &Value,
-    delay_ms: u64,
-) -> Result<Value, String> {
-    if !advance_job_progress(state, job_id, 20, "collecting changed files", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let changed_files = strings_from_array(
-        arguments
-            .get("changed_files")
-            .and_then(|value| value.as_array()),
-        "file",
-        8,
-    );
-    let target_files = if !changed_files.is_empty() {
-        changed_files
-    } else if let Some(path) = arguments.get("path").and_then(|value| value.as_str()) {
-        vec![path.to_owned()]
-    } else {
-        let changed =
-            super::graph::get_changed_files_tool(state, &json!({"include_untracked": true}))
-                .map(|out| out.0)
-                .unwrap_or_else(|_| json!({"files": [], "count": 0}));
-        strings_from_array(
-            changed.get("files").and_then(|value| value.as_array()),
-            "file",
-            8,
-        )
-    };
-    if !advance_job_progress(state, job_id, 45, "measuring impact surface", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let mut impact_rows = Vec::new();
-    let mut top_findings = Vec::new();
-    let total = target_files.iter().take(5).count().max(1);
-    for (idx, path) in target_files.iter().take(5).enumerate() {
-        let impact = super::graph::get_impact_analysis(
-            state,
-            &json!({"file_path": path, "max_depth": 2}),
-        )
-        .map(|output| output.0)
-        .unwrap_or_else(
-            |_| json!({"file_path": path, "total_affected_files": 0, "direct_importers": []}),
-        );
-        let affected = impact
-            .get("total_affected_files")
-            .and_then(|value| value.as_u64())
-            .unwrap_or_default();
-        top_findings.push(format!("{path}: {affected} affected file(s)"));
-        impact_rows.push(json!({
-            "path": path,
-            "affected_files": affected,
-            "direct_importers": impact.get("direct_importers").cloned().unwrap_or(json!([])),
-            "blast_radius": impact.get("blast_radius").cloned().unwrap_or(json!([])),
-        }));
-        let loop_progress = 45 + (((idx + 1) * 35) / total) as u8;
-        if !advance_job_progress(
-            state,
-            job_id,
-            loop_progress.min(80),
-            &format!("analyzed impact for {path}"),
-            delay_ms,
-        )? {
-            return Ok(json!({}));
-        }
-    }
-    let mut sections = BTreeMap::new();
-    sections.insert(
-        "impact_rows".to_owned(),
-        json!({"files": target_files, "impacts": impact_rows}),
-    );
-    if !advance_job_progress(state, job_id, 90, "writing impact analysis", delay_ms)? {
-        return Ok(json!({}));
-    }
-    make_handle_response(
-        state,
-        "impact_report",
-        stable_cache_key("impact_report", arguments, &["path", "changed_files"]),
-        "Diff-aware impact report with bounded blast radius and importer evidence.".to_owned(),
-        top_findings,
-        0.88,
-        vec!["Expand only the highest-impact file before deeper review".to_owned()],
-        sections,
-        target_files,
-        None,
-    )
-    .map(|(payload, _meta)| payload)
-    .map_err(|error| error.to_string())
-}
-
-fn run_refactor_safety_report_job(
-    state: &AppState,
-    job_id: &str,
-    arguments: &Value,
-    delay_ms: u64,
-) -> Result<Value, String> {
-    let path = arguments
-        .get("path")
-        .and_then(|value| value.as_str())
-        .unwrap_or(".");
-    let task = arguments.get("task").and_then(|value| value.as_str());
-    let symbol = arguments.get("symbol").and_then(|value| value.as_str());
-    if !advance_job_progress(state, job_id, 20, "analyzing module boundaries", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let boundary = module_boundary_report(state, &json!({"path": path}))
-        .map(|output| output.0)
-        .map_err(|error| error.to_string())?;
-    if !advance_job_progress(state, job_id, 40, "summarizing symbol impact", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let symbol_impact = if let Some(symbol) = symbol {
-        summarize_symbol_impact(
-            state,
-            &json!({"symbol": symbol, "file_path": arguments.get("file_path").and_then(|v| v.as_str())}),
-        )
-        .map(|output| output.0)
-        .unwrap_or_else(|error| json!({"symbol": symbol, "error": error.to_string()}))
-    } else {
-        json!({"skipped": true, "reason": "no symbol provided"})
-    };
-    if !advance_job_progress(state, job_id, 60, "ranking refactor context", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let change_request = task
-        .map(|task| analyze_change_request(state, &json!({"task": task})).map(|output| output.0))
-        .transpose()
-        .map_err(|error| error.to_string())?
-        .unwrap_or_else(|| json!({"skipped": true, "reason": "no task provided"}));
-    if !advance_job_progress(state, job_id, 80, "collecting related tests", delay_ms)? {
-        return Ok(json!({}));
-    }
-    let tests = super::filesystem::find_tests(state, &json!({"path": path, "max_results": 10}))
-        .map(|output| output.0)
-        .unwrap_or_else(|_| json!({"tests": []}));
-
-    let mut top_findings = Vec::new();
-    if let Some(symbol) = symbol {
-        top_findings.push(format!(
-            "Validate symbol-level callers before refactoring `{symbol}`."
-        ));
-    }
-    if let Some(task) = task {
-        top_findings.push(format!("Keep the refactor aligned with `{task}`."));
-    }
-    top_findings.push(format!(
-        "Check tests around `{path}` before applying broad edits."
-    ));
-
-    let mut sections = BTreeMap::new();
-    sections.insert("module_boundary".to_owned(), boundary);
-    sections.insert("symbol_impact".to_owned(), symbol_impact);
-    sections.insert("change_request".to_owned(), change_request);
-    sections.insert("related_tests".to_owned(), tests);
-    if !advance_job_progress(
-        state,
-        job_id,
-        92,
-        "writing refactor safety analysis",
-        delay_ms,
-    )? {
-        return Ok(json!({}));
-    }
-    make_handle_response(
-        state,
-        "refactor_safety_report",
-        stable_cache_key(
-            "refactor_safety_report",
-            arguments,
-            &["task", "symbol", "path", "file_path"],
-        ),
-        format!("Preview-first refactor safety report for `{path}`."),
-        top_findings,
-        0.9,
-        vec!["Use safe_rename_report or focused edits only after checking blockers".to_owned()],
-        sections,
-        vec![
-            arguments
-                .get("file_path")
-                .and_then(|value| value.as_str())
-                .unwrap_or(path)
-                .to_owned(),
-        ],
-        symbol.map(ToOwned::to_owned),
-    )
-    .map(|(payload, _meta)| payload)
-    .map_err(|error| error.to_string())
-}
-
 pub fn analyze_change_request(state: &AppState, arguments: &Value) -> ToolResult {
     let task = required_string(arguments, "task")?;
     let ranked = super::symbols::get_ranked_context(
@@ -1293,219 +841,4 @@ pub fn diff_aware_references(state: &AppState, arguments: &Value) -> ToolResult 
         changed_files,
         None,
     )
-}
-
-pub(crate) fn run_analysis_job_from_queue(
-    worker_state: &AppState,
-    job_id: String,
-    kind: String,
-    arguments: Value,
-) -> &'static str {
-    let project_path = worker_state
-        .project()
-        .as_path()
-        .to_string_lossy()
-        .to_string();
-    if worker_state
-        .get_analysis_job(&job_id)
-        .as_ref()
-        .map(|job| job.status.as_str())
-        == Some("cancelled")
-    {
-        return "cancelled";
-    }
-    patch_job_file(
-        &project_path,
-        &job_id,
-        Some("running"),
-        Some(5),
-        Some(Some("worker started".to_owned())),
-        None,
-        None,
-    );
-    let worker = std::panic::catch_unwind(std::panic::AssertUnwindSafe(
-        || -> Result<&'static str, String> {
-            if worker_state
-                .get_analysis_job(&job_id)
-                .as_ref()
-                .map(|job| job.status.as_str())
-                == Some("cancelled")
-            {
-                return Ok("cancelled");
-            }
-            let result = run_job_kind_with_progress(worker_state, &job_id, &kind, &arguments);
-            match result {
-                Ok(payload) if payload.is_object() => {
-                    let (analysis_id, estimated_sections) = extract_handle_fields(&payload);
-                    let current = worker_state.get_analysis_job(&job_id);
-                    if current.as_ref().map(|job| job.status.as_str()) == Some("cancelled") {
-                        return Ok("cancelled");
-                    }
-                    worker_state
-                        .update_analysis_job(
-                            &job_id,
-                            Some("completed"),
-                            Some(100),
-                            Some(Some("completed".to_owned())),
-                            Some(estimated_sections),
-                            Some(analysis_id),
-                            Some(None),
-                        )
-                        .map_err(|error| error.to_string())?;
-                    Ok("completed")
-                }
-                Ok(_) => Ok("failed"),
-                Err(error) => {
-                    worker_state
-                        .update_analysis_job(
-                            &job_id,
-                            Some("failed"),
-                            Some(100),
-                            Some(Some("failed".to_owned())),
-                            None,
-                            Some(None),
-                            Some(Some(error.to_string())),
-                        )
-                        .map_err(|error| error.to_string())?;
-                    Ok("failed")
-                }
-            }
-        },
-    ));
-    match worker {
-        Err(panic) => {
-            let message = if let Some(text) = panic.downcast_ref::<&str>() {
-                (*text).to_owned()
-            } else if let Some(text) = panic.downcast_ref::<String>() {
-                text.clone()
-            } else {
-                "analysis worker panicked".to_owned()
-            };
-            patch_job_file(
-                &project_path,
-                &job_id,
-                Some("failed"),
-                Some(100),
-                Some(Some("failed".to_owned())),
-                Some(None),
-                Some(Some(message)),
-            );
-            "failed"
-        }
-        Ok(Err(message)) => {
-            patch_job_file(
-                &project_path,
-                &job_id,
-                Some("failed"),
-                Some(100),
-                Some(Some("failed".to_owned())),
-                Some(None),
-                Some(Some(message)),
-            );
-            "failed"
-        }
-        Ok(Ok(status)) => status,
-    }
-}
-
-pub fn start_analysis_job(state: &AppState, arguments: &Value) -> ToolResult {
-    let kind = required_string(arguments, "kind")?.to_owned();
-    let profile_hint = arguments
-        .get("profile_hint")
-        .and_then(|value| value.as_str())
-        .map(ToOwned::to_owned);
-    let estimated_sections = estimated_sections_for_kind(&kind);
-    let job = state.store_analysis_job(
-        &kind,
-        profile_hint.clone(),
-        estimated_sections.clone(),
-        "queued",
-        0,
-        Some("queued".to_owned()),
-        None,
-        None,
-    )?;
-    state.enqueue_analysis_job(
-        job.id.clone(),
-        kind,
-        arguments.clone(),
-        profile_hint.clone(),
-    )?;
-    Ok((
-        json!({
-            "job_id": job.id,
-            "status": job.status,
-            "progress": job.progress,
-            "current_step": job.current_step,
-            "analysis_id": job.analysis_id,
-            "estimated_sections": estimated_sections,
-        }),
-        success_meta(BackendKind::Hybrid, 0.92),
-    ))
-}
-
-pub fn get_analysis_job(state: &AppState, arguments: &Value) -> ToolResult {
-    let job_id = required_string(arguments, "job_id")?;
-    let job = state
-        .get_analysis_job(job_id)
-        .ok_or_else(|| CodeLensError::NotFound(format!("unknown job_id `{job_id}`")))?;
-    Ok((
-        json!({
-            "job_id": job.id,
-            "kind": job.kind,
-            "status": job.status,
-            "progress": job.progress,
-            "current_step": job.current_step,
-            "profile_hint": job.profile_hint,
-            "estimated_sections": job.estimated_sections,
-            "analysis_id": job.analysis_id,
-            "error": job.error,
-            "updated_at_ms": job.updated_at_ms,
-        }),
-        success_meta(BackendKind::Memory, 1.0),
-    ))
-}
-
-pub fn cancel_analysis_job(state: &AppState, arguments: &Value) -> ToolResult {
-    let job_id = required_string(arguments, "job_id")?;
-    let job = state.cancel_analysis_job(job_id)?;
-    Ok((
-        json!({
-            "job_id": job.id,
-            "status": job.status,
-            "progress": job.progress,
-            "current_step": job.current_step,
-            "analysis_id": job.analysis_id,
-        }),
-        success_meta(BackendKind::Memory, 1.0),
-    ))
-}
-
-pub fn get_analysis_section(state: &AppState, arguments: &Value) -> ToolResult {
-    let analysis_id = required_string(arguments, "analysis_id")?;
-    let section = required_string(arguments, "section")?;
-    let artifact = state
-        .get_analysis(analysis_id)
-        .ok_or_else(|| CodeLensError::NotFound(format!("unknown analysis_id `{analysis_id}`")))?;
-    let content =
-        state
-            .get_analysis_section(analysis_id, section)
-            .map_err(|error| match error {
-                CodeLensError::Io(io) if io.kind() == std::io::ErrorKind::NotFound => {
-                    CodeLensError::NotFound(format!(
-                        "analysis `{analysis_id}` has no section `{section}`"
-                    ))
-                }
-                other => other,
-            })?;
-    Ok((
-        json!({
-            "analysis_id": analysis_id,
-            "section": section,
-            "content": content,
-            "tool_name": artifact.tool_name,
-            "surface": artifact.surface,
-        }),
-        success_meta(BackendKind::Memory, artifact.confidence),
-    ))
 }
